added DevToken to Client
